separate out PermutationGenerator
Move PermutationGenerator to its own package for use in other
packages, such as the new imagine tool.
